Scramble credentials and use separate discoverable IV
This fixes exactly one FIDO test case: where a valid RK is deleted,
and then its credential ID is passed to the authenticator in an
allowList (as if it were a non-discoverable credential).

This test case is fixed at great cost by using a separate IV for
discoverable and non-discoverable credentials, and in event we're
given a discoverable credential ID in the allowList, we check it's
still present in our resident keys before using it.

In effect, deleting a discoverable credential now truly disables it.

Additionally, the credential ID encoding is improved by mixing the
private key and RP ID hash components together, reducing the influence
the choice of RP ID has on the output credential ID blob.